requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
joblib==1.5.2
numpy==2.2.6
scikit-learn==1.7.2
scipy==1.15.3
sqlparse==0.5.3
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
